Latest Server Version
The latest version from the main server.
Moved script loading out so they can be kept on Dropbox etc. This reflect this in the file structure change.

Added TemplateLoad.lua
Sleepcycle.lua

changes to
fobhandler - updated a set of code that checks if a fob is red or blue to read the actual static now. Also do a check at the time of spawn to make certain that we are still held by said group. (and hope it doesn't break anything).

Slothandler - adjusted the ctld spawn in to be 60 seconds after and to check if the base is still owned by the same side as the requested spawn.

Fixed a bug that was causing 2 of every ctld item to spawn in